requirements.txt
flask
openai-whisper
pyaudio
flask-socketio
eventlet